Move GenerateKeyType out of matcher; Create derivation module
